+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Parentheses are not necessary anymore
 - improved comments of DidCommit, DidPublish and DidPush
 - Improved help for CanCommit, CanPush, CanPublish
+- DidUpgrade checks also if everything is upgraded
 
 ### Removed